move some code around and add some comment for better readability
